feat(portfolio): unify portfolio services and constants, remove unused mock data
