+++ src/services/finances/index.ts
@@ -1,25 +1,19 @@
-import { ENV } from 'src/consts';
-import { financialPlansService } from './personalFinancesService';
-import { mockFinancialPlansService } from './mockFinancialPlansService';
-import { portfolioService, mockPortfolioService } from './portfolioService';
-
-export const createFinancialPlansService = (forceMock: boolean = false) => {
-  return ENV.USE_MOCKED_DATA || forceMock
-    ? mockFinancialPlansService
-    : financialPlansService;
-};
-
-export const createPortfolioService = (forceMock: boolean = false) => {
-  return ENV.USE_MOCKED_DATA || forceMock
-    ? mockPortfolioService
-    : portfolioService;
-};
+import {
+  financialPlansService,
+  mockFinancialPlansService,
+  createFinancialPlansService,
+} from './personalFinancesService';
+import {
+  portfolioService,
+  mockPortfolioService,
+  createPortfolioService,
+} from './portfolioService';
 
 export {
   financialPlansService,
   mockFinancialPlansService,
+  createFinancialPlansService,
   portfolioService,
   mockPortfolioService,
+  createPortfolioService,
 };
-
-export const defaultFinancialPlansService = createFinancialPlansService();
